refactor: variants 추상화
